Move sortFunctions.js module into systems/renderer.js
